WIP: Sandbox implementation.
Current state:
- I'm able to send messages to the sandbox.
- The sandbox is able to render the Handlebars content.
- *The sandbox cannot send messages back to the caller!*

I don't understand why this isn't working given that I'm following the
doc here:
https://developer.chrome.com/docs/extensions/mv2/sandboxingEval/; when I
attempt to send a message to the source of my message, the source is the
sandboxed frame.  This is obviously not going to work given that I want
to send the message to the parent frame instead (& unforch, you can't
access `window.parent` either).

Definitely perplexed and I've already spent enough time on this today.